# test/e2e/__snapshots__/test_openapi/test_OpenAPIの形が変わっていないことを確認.json
{
  "components": {
    "schemas": {
      "AccentPhrase": {
        "description": "アクセント句ごとの情報。",
        "properties": {
          "accent": {
            "description": "アクセント箇所",
            "title": "Accent",
            "type": "integer"
          },
          "is_interrogative": {
            "default": false,
            "description": "疑問系かどうか",
            "title": "Is Interrogative",
            "type": "boolean"
          },
          "moras": {
            "description": "モーラのリスト",
            "items": {
              "$ref": "#/components/schemas/Mora"
            },
            "title": "Moras",
            "type": "array"
          },
          "pause_mora": {
            "$ref": "#/components/schemas/Mora",
            "description": "アクセント句の末尾につく無音モーラ。null の場合は無音モーラを付けない。",
            "title": "Pause Mora"
          }
        },
        "required": [
          "moras",
          "accent"
        ],
        "title": "AccentPhrase",
        "type": "object"
      },
...
      "AudioQuery": {
        "description": "音声合成用のクエリ。",
        "properties": {
          "accent_phrases": {
            "description": "アクセント句のリスト。",
            "items": {
              "$ref": "#/components/schemas/AccentPhrase"
            },
            "title": "Accent Phrases",
            "type": "array"
          },
          "intonationScale": {
            "description": "選択した話者スタイルの感情表現の強弱を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n「全体の抑揚」ではない点で VOICEVOX ENGINE と仕様が異なる。\n数値が大きいほど、選択した話者スタイルに近い感情表現が込められた声になる。\n例えば話者スタイルが「上機嫌」なら、数値が大きいほどより嬉しそうな明るい話し方になる。\n一方で、話者やスタイルによっては、数値を上げすぎると発声がおかしくなったり、棒読みで不自然な声になる場合もある。\n正しく発声できる上限値は話者やスタイルごとに異なる。必要に応じて最適な値を見つけて調整すること。\n全スタイルの平均であるノーマルスタイルでは自動で適切な感情表現が選択されるため、この値を指定しても無視される。",
            "title": "Intonationscale",
            "type": "number"
          },
          "kana": {
            "description": "読み上げるテキストを指定する。「読みの AquesTalk 風記法テキスト」ではない点で VOICEVOX ENGINE と仕様が異なる。\nVOICEVOX ENGINE では AquesTalk 風記法テキストが入る読み取り専用フィールドだが (音声合成時には無視される) 、AivisSpeech Engine では音声合成時に漢字や記号が含まれた通常の読み上げテキストも必要なため、苦肉の策で読み上げテキスト指定用のフィールドとして転用した。\nVOICEVOX ENGINE との互換性のため None や空文字列が指定された場合も動作するが、その場合はアクセント句から自動生成されたひらがな文字列が読み上げテキストになるため、不自然なイントネーションになってしまう。\n可能な限り kana に通常の読み上げテキストを指定した上で音声合成 API に渡すことを推奨する。",
            "title": "Kana",
            "type": "string"
          },
          "outputSamplingRate": {
            "description": "音声データの出力サンプリングレート。",
            "title": "Outputsamplingrate",
            "type": "integer"
          },
          "outputStereo": {
            "description": "音声データをステレオ出力するか否か。",
            "title": "Outputstereo",
            "type": "boolean"
          },
          "pauseLength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "句読点などの無音時間。null のときは無視される。デフォルト値は null 。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)"
          },
          "pauseLengthScale": {
            "default": 1,
            "description": "句読点などの無音時間（倍率）。デフォルト値は 1 。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)",
            "type": "number"
          },
          "pitchScale": {
            "description": "全体の音高を -0.15 ~ 0.15 の範囲で指定する (デフォルト: 0.0) 。\n数値が大きいほど高い声になる。\nVOICEVOX ENGINE と異なり、この値を 0.0 から変更すると音質が劣化するため注意が必要。",
            "title": "Pitchscale",
            "type": "number"
          },
          "postPhonemeLength": {
            "description": "音声の後の無音時間 (秒)。",
            "title": "Postphonemelength",
            "type": "number"
          },
          "prePhonemeLength": {
            "description": "音声の前の無音時間 (秒)。",
            "title": "Prephonemelength",
            "type": "number"
          },
          "speedScale": {
            "description": "全体の話速を 0.5 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n2.0 で 2 倍速、0.5 で 0.5 倍速になる。",
            "title": "Speedscale",
            "type": "number"
          },
          "tempoDynamicsScale": {
            "default": 1.0,
            "description": "話す速さ（テンポ）の緩急の強弱を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\nAivisSpeech Engine 固有のフィールドで、VOICEVOX ENGINE には存在しない。\n数値が大きいほどより早口で生っぽい抑揚がついた声になる。\nVOICEVOX ENGINE との互換性のため、未指定時はデフォルト値が適用される。",
            "title": "Tempodynamicsscale",
            "type": "number"
          },
          "volumeScale": {
            "description": "全体の音量を 0.0 ~ 2.0 の範囲で指定する (デフォルト: 1.0) 。\n数値が大きいほど大きな声になる。",
            "title": "Volumescale",
            "type": "number"
          }
        },
        "required": [
          "accent_phrases",
          "speedScale",
          "intonationScale",
          "pitchScale",
          "volumeScale",
          "prePhonemeLength",
          "postPhonemeLength",
          "outputSamplingRate",
          "outputStereo"
        ],
        "title": "AudioQuery",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Mora": {
        "description": "モーラ（子音＋母音）ごとの情報。",
        "properties": {
          "consonant": {
            "description": "子音の音素。",
            "title": "Consonant",
            "type": "string"
          },
          "consonant_length": {
            "description": "子音の音長。\nAivisSpeech Engine の実装上算出できないため、ダミー値として常に 0.0 が返される。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)",
            "type": "number"
          },
          "pitch": {
            "description": "音高。\nAivisSpeech Engine の実装上算出できないため、ダミー値として常に 0.0 が返される。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)",
            "type": "number"
          },
          "text": {
            "description": "子音＋母音に対応する文字。\nVOICEVOX ENGINE と異なり、感嘆符・句読点などの記号もモーラに含まれる。\n記号モーラの場合、`text` には記号がそのまま、`vowel` には \"pau\" が設定される。",
            "title": "Text",
            "type": "string"
          },
          "vowel": {
            "description": "母音の音素。",
            "title": "Vowel",
            "type": "string"
          },
          "vowel_length": {
            "description": "母音の音長。\nAivisSpeech Engine の実装上算出できないため、ダミー値として常に 0.0 が返される。",
            "title": "AivisSpeech Engine ではサポートされていないフィールドです (常に無視されます)",
            "type": "number"
          }
        },
        "required": [
          "text",
          "vowel",
          "vowel_length",
          "pitch"
        ],
        "title": "Mora",
        "type": "object"
      },
      "MorphableTargetInfo": {
        "description": "モーフィング相手としての情報。",
        "properties": {
          "is_morphable": {
            "description": "指定したキャラクターに対してモーフィングの可否",
            "title": "Is Morphable",
            "type": "boolean"
          }
        },
        "required": [
          "is_morphable"
        ],
        "title": "MorphableTargetInfo",
        "type": "object"
      },
...
      "Speaker": {
        "description": "キャラクター情報",
        "properties": {
          "name": {
            "description": "名前",
            "title": "Name",
            "type": "string"
          },
          "speaker_uuid": {
            "description": "キャラクターの UUID",
            "title": "Speaker Uuid",
            "type": "string"
          },
          "styles": {
            "description": "スタイルの一覧",
            "items": {
              "$ref": "#/components/schemas/SpeakerStyle"
            },
            "title": "Styles",
            "type": "array"
          },
          "supported_features": {
            "$ref": "#/components/schemas/SpeakerSupportedFeatures",
            "description": "キャラクターの対応機能"
          },
          "version": {
            "description": "キャラクターのバージョン",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "speaker_uuid",
          "styles",
          "version"
        ],
        "title": "Speaker",
        "type": "object"
      },
      "SpeakerInfo": {
        "description": "キャラクターの追加情報",
        "properties": {
          "policy": {
            "description": "policy.md",
            "title": "Policy",
            "type": "string"
          },
          "portrait": {
            "description": "アイコンを base64 エンコードしたもの、あるいは URL (「立ち絵画像を base64 エンコードしたもの」ではない点で VOICEVOX ENGINE と異なる)",
            "title": "Portrait",
            "type": "string"
          },
          "style_infos": {
            "description": "スタイルの追加情報",
            "items": {
              "$ref": "#/components/schemas/StyleInfo"
            },
            "title": "Style Infos",
            "type": "array"
          }
        },
        "required": [
          "policy",
          "portrait",
          "style_infos"
        ],
        "title": "SpeakerInfo",
        "type": "object"
      },
      "SpeakerStyle": {
        "description": "キャラクターのスタイル情報",
        "properties": {
          "id": {
            "description": "スタイル ID",
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "description": "スタイル名",
            "title": "Name",
            "type": "string"
          },
          "type": {
            "default": "talk",
            "description": "スタイルの種類。talk: 音声合成クエリの作成と音声合成が可能。singing_teacher: 歌唱音声合成用のクエリの作成が可能。frame_decode: 歌唱音声合成が可能。sing: 歌唱音声合成用のクエリの作成と歌唱音声合成が可能。",
            "enum": [
              "talk",
              "singing_teacher",
              "frame_decode",
              "sing"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "SpeakerStyle",
        "type": "object"
      },
      "SpeakerSupportedFeatures": {
        "description": "キャラクターの対応機能の情報",
        "properties": {
          "permitted_synthesis_morphing": {
            "default": "ALL",
            "description": "モーフィング機能への対応。'ALL' は「全て許可」、'SELF_ONLY' は「同じキャラクター内でのみ許可」、'NOTHING' は「全て禁止」",
            "enum": [
              "ALL",
              "SELF_ONLY",
              "NOTHING"
            ],
            "title": "Permitted Synthesis Morphing",
            "type": "string"
          }
        },
        "title": "SpeakerSupportedFeatures",
        "type": "object"
      },
      "StyleInfo": {
        "description": "スタイルの追加情報",
        "properties": {
          "icon": {
            "description": "このスタイルのアイコンを base64 エンコードしたもの、あるいは URL",
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "description": "スタイル ID",
            "title": "Id",
            "type": "integer"
          },
          "portrait": {
            "description": "AivisSpeech Engine では常に None を返す (「このスタイルの立ち絵画像を base64 エンコードしたもの」ではない点で VOICEVOX ENGINE と異なる)",
            "title": "Portrait",
            "type": "string"
          },
          "voice_sample_transcripts": {
            "default": [],
            "description": "ボイスサンプルの書き起こしテキスト (voice_samples の配列インデックスと対応し、存在しない場合は空文字列)",
            "items": {
              "type": "string"
            },
            "title": "Voice Sample Transcripts",
            "type": "array"
          },
          "voice_samples": {
            "description": "ボイスサンプルの音声データを base64 エンコードしたもの、あるいは URL",
            "items": {
              "type": "string"
            },
            "title": "Voice Samples",
            "type": "array"
          }
        },
        "required": [
          "id",
          "icon",
          "voice_samples"
        ],
        "title": "StyleInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AivisSpeech の音声合成エンジンです。",
    "title": "AivisSpeech Engine",
    "version": "latest"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/morphable_targets": {
      "post": {
        "description": "指定されたベーススタイルに対してエンジン内の各キャラクターがモーフィング機能を利用可能か返します。<br>\nモーフィングの許可/禁止は `/speakers `の `speaker.supported_features.synthesis_morphing` に記載されています。<br>\nプロパティが存在しない場合は、モーフィングが許可されているとみなします。<br>\n返り値のスタイル ID は string 型なので注意。<br>\nAivisSpeech Engine では話者ごとに発声タイミングが異なる関係で実装不可能なため (動作こそするが聴くに耐えない) 、\n全ての話者でモーフィングが禁止されています。",
        "operationId": "morphable_targets",
        "parameters": [
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "integer"
                },
                "title": "Base Style Ids",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "$ref": "#/components/schemas/MorphableTargetInfo"
                    },
                    "type": "object"
                  },
                  "title": "Response Morphable Targets Morphable Targets Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "指定したスタイルに対してエンジン内のキャラクターがモーフィングが可能か判定する",
        "tags": [
          "音声合成"
        ]
      }
    },
...
    "/singer_info": {
      "get": {
        "operationId": "singer_info",
        "parameters": [
          {
            "in": "query",
            "name": "speaker_uuid",
            "required": true,
            "schema": {
              "title": "Speaker Uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resource_format",
            "required": false,
            "schema": {
              "default": "base64",
              "enum": [
                "base64",
                "url"
              ],
              "title": "Resource Format",
              "type": "string"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AivisSpeech Engine ではサポートされていない API です (常に 501 Not Implemented を返します)",
        "tags": [
          "話者情報"
        ]
      }
    },
    "/singers": {
      "get": {
        "operationId": "singers",
        "parameters": [
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Singers Singers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AivisSpeech Engine ではサポートされていない API です (常に 501 Not Implemented を返します)",
        "tags": [
          "話者情報"
        ]
      }
    },
    "/speaker_info": {
      "get": {
        "description": "UUID で指定された話者の情報を返します。\n画像や音声は resource_format で指定した形式で返されます。",
        "operationId": "speaker_info",
        "parameters": [
          {
            "in": "query",
            "name": "speaker_uuid",
            "required": true,
            "schema": {
              "title": "Speaker Uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resource_format",
            "required": false,
            "schema": {
              "default": "base64",
              "enum": [
                "base64",
                "url"
              ],
              "title": "Resource Format",
              "type": "string"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "UUID で指定された話者の情報を取得する",
        "tags": [
          "話者情報"
        ]
      }
    },
    "/speakers": {
      "get": {
        "description": "話者情報の一覧を返します。",
        "operationId": "speakers",
        "parameters": [
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Speaker"
                  },
                  "title": "Response Speakers Speakers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "話者情報の一覧を取得する",
        "tags": [
          "話者情報"
        ]
      }
    },
...
    "/synthesis_morphing": {
      "post": {
        "description": "指定された 2 種類のスタイルで音声を合成、指定した割合でモーフィングした音声を得ます。<br>\nモーフィングの割合は `morph_rate` で指定でき、0.0 でベースのスタイル、1.0 でターゲットのスタイルに近づきます。<br>\nAivisSpeech Engine では話者ごとに発声タイミングが異なる関係で実装不可能なため (動作こそするが聴くに耐えない) 、\n常に 400 Bad Request を返します。",
        "operationId": "_synthesis_morphing",
        "parameters": [
          {
            "in": "query",
            "name": "base_speaker",
            "required": true,
            "schema": {
              "title": "Base Speaker",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "target_speaker",
            "required": true,
            "schema": {
              "title": "Target Speaker",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "morph_rate",
            "required": true,
            "schema": {
              "maximum": 1.0,
              "minimum": 0.0,
              "title": "Morph Rate",
              "type": "number"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "enable_interrogative_upspeak",
            "required": false,
            "schema": {
              "default": true,
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Enable Interrogative Upspeak",
              "type": "boolean"
            }
          },
          {
            "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
            "in": "query",
            "name": "core_version",
            "required": false,
            "schema": {
              "description": "AivisSpeech Engine ではサポートされていないパラメータです (常に無視されます) 。",
              "title": "Core Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "2種類のスタイルでモーフィングした音声を合成する",
        "tags": [
          "音声合成"
        ]
      }
    },
...
  }
}